Deploy script that automates download
This script can be distributed separately and is used to download,
install and configure the prebuilt programs for use with OpTiMSoC.

The script is used as follows:

 optimsoc-prebuilt-deploy.sh <options> <prebuilt-list>

where options are:

 -d  Destination folder (default: /opt/optimsoc)
 -r  Ubuntu release (default: 14.04)

prebuilt-list is a list of any of:

 or1kelf: Baremetal toolchain
 verilator: Verilator (verilog simulator)
 systemc: SystemC (simulation language and library)

Currently ubuntu release 12.04 and 14.04 are used to distinguish
between different versions of libmpfr, libgmp, etc.

Between downloading and extracting the prebuilt programs, the script
also generates the script that sets bash environment variables, that
is sourced as:

 source /opt/optimsoc/setup_prebuilt.sh

or with the different destination folder.